+++ Administrador/prompt_main.py
@@ -1,15 +1,18 @@
-from turmas.ctrl_turmas import Turmas
-from times.prompt_times import ctrl_times
+from turmas import prompt_turmas
+from times import prompt_times
 import os
 
-def ctrl_turmas():
+
+
+def promptMain():
     while True:
+        os.system('cls' if os.name == 'nt' else 'clear')
         print('')
-        print("Controle de Turmas!!!")
+        print("Bem vindo!!!")
         print('')
         while True:
             try:
-                print("Escolha uma opção: \n1 - Criar nova turma\n2 - Visualizar turmas\n3 - Editar turma\n4 - Excluir turma\n5 - Controle de times\n0 - Voltar\n")
+                print("Escolha uma opção: \n1 - Controle de turmas\n2 - Controle de times\n3 - Controle de Usuários\n4 - Controle de Sprints\n0 - Voltar\n")
                 op = int(input("Digite aqui: "))
                 break
             except ValueError:
@@ -17,27 +20,21 @@ def ctrl_turmas():
         
         if op == 1:
             os.system('cls' if os.name == 'nt' else 'clear')
-            print("\nCriar nova turma!!!")
-            print(Turmas.createTurmas())
+            prompt_turmas.ctrl_turmas()
             
         elif op == 2:
-            
-            print("\nVisualizando turmas!!!")
-            Turmas.listAllTurmas()
+            os.system('cls' if os.name == 'nt' else 'clear')
+            prompt_times.ctrl_times()
             
         elif op == 3:
             os.system('cls' if os.name == 'nt' else 'clear')
-            print("\nEditar turma!!!\n")
-            print(Turmas.editTurmas())
+            #AQUI ENTRA O CONTROLE DE USUÁRIOS
+            pass
             
         elif op == 4:
             os.system('cls' if os.name == 'nt' else 'clear')
-            print("\nExcluír turma!!!")
-            print(Turmas.delTurma())
-        
-        elif op == 5:
-            os.system('cls' if os.name == 'nt' else 'clear')
-            ctrl_times()
+            #AQUI ENTRA O CONTROLE DE USUÁRIOS
+            pass
         
         elif op == 0:
             os.system('cls' if os.name == 'nt' else 'clear')
@@ -46,5 +43,5 @@ def ctrl_turmas():
         else:
             print('\nOpção inválida! Tente novamente!')
         print("----------------------------------------------------")
-
-ctrl_turmas()
+        
+promptMain()
